+++ src/front/melody-front/src/app/(board)/help/list/page.js
@@ -1,0 +1,17 @@
+"use client"
+import React, { useEffect, useState } from "react";
+import axios from "axios";
+import Link from "next/link"
+
+const page = () => {
+
+
+    return (
+        <div className="container mx-auto ">
+
+        </div>
+    );
+};
+
+export default page;
+
