--- src/front/melody-front/src/app/(board)/help/page.js
@@ -1,13 +1,0 @@
-"use client"
-
-import Link from "next/link";
-
-const ServiceGu = (children) => {
-    return (
-        <div>
-            <h1>service</h1>
-            <Link href="/service">create</Link>
-        </div>
-    )
-}
-export default ServiceGu;
