+++ src/front/melody-front/src/app/(board)/help/write/layout.js
@@ -1,0 +1,14 @@
+"use client"
+import axios from "axios";
+import Link from "next/link"
+
+const WriteLayout = () => {
+
+
+    return (
+        <div className="container mx-auto ">
+
+        </div>    );
+};
+
+export default WriteLayout;
